역할 : ArticleRepository / 구현 : MemoryArticleRepository 분리
- 데이터베이스 연동의 확장성을 위해 ArticleRepository 인터페이스를 따로 분리함
- 유효한 게시글인지 여부의 검증은 ArticleController 단에서 하도록 함
- 고로, ArticleRepository의 검증 로직이 제대로 수행되는지에 대한 테스트는 제거했다.
