+++ src/main/java/com/ttasjwi/cafe/controller/HomeController.java
@@ -2,8 +2,6 @@ package com.ttasjwi.cafe.controller;
 
 import com.ttasjwi.cafe.domain.Article;
 import com.ttasjwi.cafe.repository.ArticleRepository;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
