+++ src/test/java/com/ttasjwi/cafe/controller/HomeControllerTest.java
@@ -1,6 +1,7 @@
 package com.ttasjwi.cafe.controller;
 
 import com.ttasjwi.cafe.repository.ArticleRepository;
+import com.ttasjwi.cafe.repository.MemoryArticleRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
